refactoring and removing firebase_id column, creating endpoint for fetching routines for a user, handling preflight OPTIONS request, adding MapStruct, adding SQL migration to adjust table id sequence

+++ src/main/java/com/gymtrack/api/feature/routine/controller/RoutineController.java
@@ -1,11 +1,11 @@
 package com.gymtrack.api.feature.routine.controller;
 
+import com.gymtrack.api.exception.AuthenticationException;
 import com.gymtrack.api.feature.routine.model.Routine;
 import com.gymtrack.api.feature.routine.service.RoutineService;
-
+import com.gymtrack.api.feature.user.model.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,18 +18,29 @@ import java.util.List;
 public class RoutineController {
     private final RoutineService routineService;
 
-    @GetMapping
+    //    @GetMapping
     public List<Routine> getRoutines() {
         return routineService.getRoutines();
     }
 
     @GetMapping("{id}")
-    public Routine getRoutine(@PathVariable Integer id) {
-        return routineService.getRoutine(id);
+    public Routine getRoutine(@PathVariable Integer id, @RequestAttribute(value = "user") User user) {
+        Routine routine = routineService.getRoutine(id);
+        // @TODO if super-admin, can override + create validation/ACL class/layer
+        if (routine.getCreatorId().equals(user.getId())) {
+            return routine;
+        }
+
+        return null;
     }
 
     @DeleteMapping("{id}")
-    public Integer deleteRoutine(@PathVariable Integer id) {
+    public Integer deleteRoutine(@PathVariable Integer id, @RequestAttribute(value = "user") User user) {
+        Routine routine = routineService.getRoutine(id);
+        // @TODO if super-admin, can override
+        if (routine.getCreatorId().equals(user.getId())) {
+            throw new AuthenticationException();
+        }
         return routineService.deleteRoutine(id);
     }
 }
